Animate the car forward #21
