+++ src/lib/babylon/scene/animate-the-car-forward-scene.ts
